Site: created a parent Project class (/domain/project.ts) that LibraryProject extends, removed /teacher/project.ts
