+++ src/main/webapp/site/src/app/teacher/run-menu/run-menu.component.spec.ts
@@ -2,7 +2,7 @@ import { async, ComponentFixture, TestBed } from '@angular/core/testing';
 import { RunMenuComponent } from "./run-menu.component";
 import { TeacherService } from "../teacher.service";
 import { Run } from "../../domain/run";
-import { Project } from "../project";
+import { Project } from "../../domain/project";
 import { FormsModule } from "@angular/forms";
 import {
   MatDialog,
