+++ src/main/webapp/site/src/app/teacher/share-run-dialog/share-run-dialog.component.spec.ts
@@ -4,7 +4,7 @@ import { ShareRunDialogComponent } from './share-run-dialog.component';
 import { Observable } from "rxjs";
 import { MatDialogRef } from "@angular/material/dialog";
 import { MAT_DIALOG_DATA } from "@angular/material/dialog";
-import { Project } from "../project";
+import { Project } from "../../domain/project";
 import { TeacherService } from "../teacher.service";
 import { BrowserAnimationsModule } from "../../../../../../../../node_modules/@angular/platform-browser/animations";
 import { Run } from "../../domain/run";
